break components into separate files
