+++ react/typings/vtex.styleguide.d.ts
@@ -1,0 +1,5 @@
+import React from 'react'
+
+declare module 'vtex.styleguide' {
+  const IconCaretRight: React.FC<any>
+}
